requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
python-dotenv>=1.1.0
pydantic>=2.11.0
openai>=1.75.0
openai-agents>=0.2.0
openai-chatkit>=0.2.0
pinecone>=6.0.0
httpx>=0.28.0
PyYAML>=6.0.2
